+++ src/fr/maxlego08/menu/MenuPlugin.java
@@ -34,7 +34,6 @@ import fr.maxlego08.menu.zcore.enums.EnumInventory;
 import fr.maxlego08.menu.zcore.utils.nms.NMSUtils;
 import fr.maxlego08.menu.zcore.utils.plugins.Metrics;
 import fr.maxlego08.menu.zcore.utils.plugins.Plugins;
-import org.bukkit.Bukkit;
 import org.bukkit.plugin.ServicePriority;
 
 import java.io.File;